requirements.txt
wheel
gevent
flask
requests
pygments
dateutils
fuzzywuzzy
redis
colored
langdetect
cffi
polyglot
colorama
pyyaml
python-Levenshtein
